requirements.txt
pandas
turftopic
topic-wizard
datasets
transformers
spacy
iso639-lang
